+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- 修改：导出的包不再包含 test 内容。
+- 修复： execute_straight 当 shell 为 True 时，windows 使用双斜杠路径会导致执行失败，解决办法为当 shell 为 True 时，对 cmd 先做一次 normpath。
+
 # [0.2.16] 22-12-22
 
 - 修复：executor 解决 shell 为 False 的情况下如果包装了 cmd 在 windows 上会出现 ‘PermissionError: [WinError 5] 拒绝访问。’ 的问题。